Comment out frame buffer mapping.
WDDM DOD driver doesn't touch the frame buffer (bar0), so no reason to map in into memory.
Especially as there were rare BSOD failures when the mapping failed.  Leaving the code
in place so it clear where the frame buffer should be mapped if we need to do it in the future.