Calculate present buffer size without current mode
There are reported bugchecks at MmProbeAndLockPages. This may be a result
of calculating the locked buffer size with the current mode size. With
this patch, the buffer size is calculated with the pitch and rects
provided by the caller.

Also, a large portion of code has been merged between the VgaDevice and
QxlDevice classes.